backend/requirements.txt
asgiref==3.7.2
django
djangorestframework
djangorestframework-simplejwt
qrcode
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
django-crispy-forms
crispy-bootstrap4
python-dotenv
pillow
